fix(automation): relation picker resolves registered stores, no raw pb access
Adds collectionByName to core/lib/pocketbase.ts for dynamic-by-name
lookups against the pbtsdb store map, and reworks
RelationRecordPicker to use it with a plain useLiveQuery (falling back
to a plain record-id input when the target collection isn't
registered) instead of a raw pb.collection().getList() call.

Also splits ConditionsCard into ConditionGroupBox and MatchPillPair to
shrink the array-index-key surface to just the two files that
genuinely need it.
