+++ core/components/rules/ConditionGroupBox.tsx
@@ -1,0 +1,93 @@
+import type { CatalogField } from '@tinycld/core/lib/automation/api'
+import {
+    addCondition,
+    removeCondition,
+    removeGroup,
+    setGroupMatch,
+    updateCondition,
+} from '@tinycld/core/lib/automation/condition-helpers'
+import type { RuleDraft } from '@tinycld/core/lib/automation/draft'
+import { useThemeColor } from '@tinycld/core/lib/use-app-theme'
+import { Plus, Trash2 } from 'lucide-react-native'
+import { Pressable, Text, View } from 'react-native'
+import { ConditionRow } from './ConditionRow'
+import { MatchPillPair } from './MatchPillPair'
+
+export interface ConditionGroupBoxProps {
+    draft: RuleDraft
+    fields: CatalogField[]
+    groupIndex: number
+    onChange: (patch: Partial<RuleDraft>) => void
+}
+
+// Groups/conditions carry no stable id in the draft AST (see draft.ts's
+// ConditionsAstDraft) — only appended/removed via the button below, never
+// reordered, so an index key is safe here despite the usual risk of index
+// keys breaking state identity across reorders.
+export function ConditionGroupBox({ draft, fields, groupIndex, onChange }: ConditionGroupBoxProps) {
+    const mutedColor = useThemeColor('muted-foreground')
+    const group = draft.conditions.groups[groupIndex]
+
+    return (
+        <View className="rounded-lg border p-3 gap-2.5 border-border">
+            <View className="flex-row items-center justify-between">
+                <View className="flex-row items-center gap-2">
+                    <Text className="text-xs text-muted-foreground">Match</Text>
+                    <MatchPillPair
+                        match={group.match}
+                        onSelect={match =>
+                            onChange({
+                                conditions: setGroupMatch(draft.conditions, groupIndex, match),
+                            })
+                        }
+                    />
+                    <Text className="text-xs text-muted-foreground">conditions</Text>
+                </View>
+                <Pressable
+                    onPress={() =>
+                        onChange({ conditions: removeGroup(draft.conditions, groupIndex) })
+                    }
+                    className="p-1.5"
+                    hitSlop={8}
+                >
+                    <Trash2 size={14} color={mutedColor} />
+                </Pressable>
+            </View>
+
+            {group.conditions.map((condition, conditionIndex) => (
+                <ConditionRow
+                    key={conditionIndex}
+                    condition={condition}
+                    fields={fields}
+                    onChange={patch =>
+                        onChange({
+                            conditions: updateCondition(
+                                draft.conditions,
+                                groupIndex,
+                                conditionIndex,
+                                patch
+                            ),
+                        })
+                    }
+                    onRemove={() =>
+                        onChange({
+                            conditions: removeCondition(
+                                draft.conditions,
+                                groupIndex,
+                                conditionIndex
+                            ),
+                        })
+                    }
+                />
+            ))}
+
+            <Pressable
+                onPress={() => onChange({ conditions: addCondition(draft.conditions, groupIndex) })}
+                className="self-start flex-row items-center gap-1 py-1"
+            >
+                <Plus size={13} color={mutedColor} />
+                <Text className="text-xs text-muted-foreground">add condition</Text>
+            </Pressable>
+        </View>
+    )
+}
