+++ core/components/rules/RelationRecordPicker.tsx
@@ -1,8 +1,9 @@
-import { useQuery } from '@tanstack/react-query'
+import { useLiveQuery } from '@tanstack/react-db'
 import { MenuActionItem } from '@tinycld/core/components/DropdownMenu'
-import { pb } from '@tinycld/core/lib/pocketbase'
+import { collectionByName } from '@tinycld/core/lib/pocketbase'
 import { useThemeColor } from '@tinycld/core/lib/use-app-theme'
 import { Menu } from '@tinycld/core/ui/menu'
+import { PlainInput } from '@tinycld/core/ui/PlainInput'
 import { ChevronDown } from 'lucide-react-native'
 import { Pressable, Text } from 'react-native'
 
@@ -13,27 +14,35 @@ export interface RelationRecordPickerProps {
     onChange: (id: string) => void
 }
 
-type PickerRecord = { id: string } & Record<string, unknown>
-
-function recordLabel(record: PickerRecord, displayField: string): string {
+function recordLabel(record: Record<string, unknown>, displayField: string): string {
     const raw = record[displayField]
-    return typeof raw === 'string' && raw ? raw : record.id
+    const id = record.id
+    return typeof raw === 'string' && raw ? raw : typeof id === 'string' ? id : ''
 }
 
 // Relation targets are arbitrary collections declared by a package's
 // automation catalog — not guaranteed to be registered in the client's
-// pbtsdb store map (`useStore` is typed by literal collection name, so it
-// can't be called with a dynamic `target: string`). A one-shot, read-only
-// PB list fetch is the pragmatic fallback: it's allowed under the
-// "read-only reads" exception since rule authoring only needs to *display*
-// candidate records to pick from, not live-sync them.
+// pbtsdb store map (an unlinked package, or a name outside the tinycld
+// schema entirely). collectionByName resolves it dynamically when it IS
+// registered; the query itself no-ops (returns null, same pattern as
+// useOrgLiveQuery) when it isn't, rather than conditionally skipping the
+// useLiveQuery call — hooks must run unconditionally.
+//
+// Raw useLiveQuery (not useOrgLiveQuery) is correct here, same rationale as
+// use-packages.ts: the collection may be a global, non-org-scoped store (or
+// belong to another package entirely), so org/user scoping is the caller's
+// concern, not this generic picker's — rows are already RLS-filtered by the
+// server.
 function useRelationRecords(target: string) {
-    return useQuery({
-        queryKey: ['automation-relation-records', target],
-        // biome-ignore lint/plugin/pbtsdb-no-raw-pb-access: target is an arbitrary collection from a package's automation catalog, not guaranteed to be in the pbtsdb store map (useStore can't be called dynamically) — no collection object exists to hand to readCollectionCached, so a one-shot read-only network call is the pragmatic fallback.
-        queryFn: () => pb.collection(target).getList<PickerRecord>(1, 50),
-        enabled: Boolean(target),
-    })
+    const collection = collectionByName(target)
+    const { data } = useLiveQuery(
+        q => {
+            if (!collection) return null
+            return q.from({ record: collection }).limit(50)
+        },
+        [collection]
+    )
+    return { isRegistered: Boolean(collection), records: (data ?? []) as Record<string, unknown>[] }
 }
 
 export function RelationRecordPicker({
@@ -43,8 +52,19 @@ export function RelationRecordPicker({
     onChange,
 }: RelationRecordPickerProps) {
     const mutedColor = useThemeColor('muted-foreground')
-    const { data } = useRelationRecords(target)
-    const records = data?.items ?? []
+    const { isRegistered, records } = useRelationRecords(target)
+
+    if (!isRegistered) {
+        return (
+            <PlainInput
+                value={value}
+                onChangeText={onChange}
+                placeholder={`record id — ${target} isn't installed here`}
+                className="flex-1 border rounded-lg px-2.5 py-1.5 text-sm text-foreground bg-background border-border"
+            />
+        )
+    }
+
     const selected = records.find(r => r.id === value)
     const label = selected ? recordLabel(selected, displayField) : value || 'Select…'
 
@@ -63,10 +83,10 @@ export function RelationRecordPicker({
                 <Menu.Content presentation="popover" placement="bottom" align="start">
                     {records.map(record => (
                         <MenuActionItem
-                            key={record.id}
+                            key={record.id as string}
                             label={recordLabel(record, displayField)}
                             isActive={record.id === value}
-                            onPress={() => onChange(record.id)}
+                            onPress={() => onChange(record.id as string)}
                         />
                     ))}
                 </Menu.Content>
